+++ src/main/java/org/wavescale/sourcesync/action/ActionLocalFileToRemote.java
@@ -1,6 +1,14 @@
 package org.wavescale.sourcesync.action;
 
-import java.nio.file.Path;
+import com.intellij.openapi.actionSystem.AnAction;
+import com.intellij.openapi.actionSystem.AnActionEvent;
+import com.intellij.openapi.actionSystem.PlatformDataKeys;
+import com.intellij.openapi.progress.ProgressIndicator;
+import com.intellij.openapi.progress.ProgressManager;
+import com.intellij.openapi.progress.Task;
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.vfs.VirtualFile;
+import com.intellij.project.ProjectKt;
 import org.jetbrains.annotations.NotNull;
 import org.wavescale.sourcesync.api.ConnectionConfiguration;
 import org.wavescale.sourcesync.api.ConnectionConstants;
@@ -18,15 +26,8 @@ import org.wavescale.sourcesync.synchronizer.FTPFileSynchronizer;
 import org.wavescale.sourcesync.synchronizer.FTPSFileSynchronizer;
 import org.wavescale.sourcesync.synchronizer.SCPFileSynchronizer;
 import org.wavescale.sourcesync.synchronizer.SFTPFileSynchronizer;
-import com.intellij.openapi.actionSystem.AnAction;
-import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.actionSystem.PlatformDataKeys;
-import com.intellij.openapi.progress.ProgressIndicator;
-import com.intellij.openapi.progress.ProgressManager;
-import com.intellij.openapi.progress.Task;
-import com.intellij.openapi.project.Project;
-import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.project.ProjectKt;
+
+import java.nio.file.Path;
 
 /**
  * ****************************************************************************
@@ -39,25 +40,21 @@ import com.intellij.project.ProjectKt;
  * For any issues or questions send an email at: fioan89@gmail.com              *
  * *****************************************************************************
  */
-public class ActionLocalFileToRemote extends AnAction
-{
-    public void actionPerformed(final AnActionEvent e)
-    {
+public class ActionLocalFileToRemote extends AnAction {
+    public void actionPerformed(final AnActionEvent e) {
         // first check if there's a connection type associated to this module.
         // If not alert the user and get out
         final Project project = e.getProject();
 
         String projectName = project.getName();
         String associationName = ConnectionConfig.getInstance().getAssociationFor(projectName);
-        if (associationName == null)
-        {
+        if (associationName == null) {
             Utils.showNoConnectionSpecifiedError(projectName);
             return;
         }
         final VirtualFile virtualFile = PlatformDataKeys.VIRTUAL_FILE.getData(e.getDataContext());
 
-        if (virtualFile == null || virtualFile.isDirectory())
-        {
+        if (virtualFile == null || virtualFile.isDirectory()) {
             StringBuilder builder = new StringBuilder("Project <b>");
             builder.append(projectName).append("</b>! does not have a selected file!");
             BalloonLogger.logBalloonInfo(builder.toString(), project);
@@ -66,38 +63,27 @@ public class ActionLocalFileToRemote extends AnAction
         }
 
         final ConnectionConfiguration connectionConfiguration = ConfigConnectionFactory.getInstance().getConnectionConfiguration(associationName);
-        if (Utils.canBeUploaded(virtualFile.getName(), connectionConfiguration.getExcludedFiles()))
-        {
-            final Path uploadLocation = Utils.dirsToFileFromProjectRoot(virtualFile, ProjectKt.getStateStore(project));
-            ProgressManager.getInstance().run(new Task.Backgroundable(e.getProject(), "Uploading", false)
-            {
+        if (Utils.canBeUploaded(virtualFile.getName(), connectionConfiguration.getExcludedFiles())) {
+            final Path uploadLocation = Utils.relativeLocalUploadDirs(virtualFile, ProjectKt.getStateStore(project));
+            ProgressManager.getInstance().run(new Task.Backgroundable(e.getProject(), "Uploading", false) {
                 @Override
-                public void run(@NotNull ProgressIndicator indicator)
-                {
+                public void run(@NotNull ProgressIndicator indicator) {
                     FileSynchronizer fileSynchronizer = null;
-                    if (ConnectionConstants.CONN_TYPE_SCP.equals(connectionConfiguration.getConnectionType()))
-                    {
+                    if (ConnectionConstants.CONN_TYPE_SCP.equals(connectionConfiguration.getConnectionType())) {
                         fileSynchronizer = new SCPFileSynchronizer((SCPConfiguration) connectionConfiguration,
-                            project, indicator);
-                    }
-                    else if (ConnectionConstants.CONN_TYPE_SFTP.equals(connectionConfiguration.getConnectionType()))
-                    {
+                                project, indicator);
+                    } else if (ConnectionConstants.CONN_TYPE_SFTP.equals(connectionConfiguration.getConnectionType())) {
                         fileSynchronizer = new SFTPFileSynchronizer((SFTPConfiguration) connectionConfiguration,
-                            project, indicator);
-                    }
-                    else if (ConnectionConstants.CONN_TYPE_FTP.equals(connectionConfiguration.getConnectionType()))
-                    {
+                                project, indicator);
+                    } else if (ConnectionConstants.CONN_TYPE_FTP.equals(connectionConfiguration.getConnectionType())) {
                         fileSynchronizer = new FTPFileSynchronizer((FTPConfiguration) connectionConfiguration,
-                            project, indicator);
-                    }
-                    else if (ConnectionConstants.CONN_TYPE_FTPS.equals(connectionConfiguration.getConnectionType()))
-                    {
+                                project, indicator);
+                    } else if (ConnectionConstants.CONN_TYPE_FTPS.equals(connectionConfiguration.getConnectionType())) {
                         fileSynchronizer = new FTPSFileSynchronizer((FTPSConfiguration) connectionConfiguration,
-                            project, indicator);
+                                project, indicator);
                     }
 
-                    if (fileSynchronizer != null)
-                    {
+                    if (fileSynchronizer != null) {
                         fileSynchronizer.connect();
                         // so final destination will look like this:
                         // root_home/ + project_relative_path_to_file/
@@ -106,9 +92,7 @@ public class ActionLocalFileToRemote extends AnAction
                     }
                 }
             });
-        }
-        else
-        {
+        } else {
             EventDataLogger.logWarning("File <b>" + virtualFile.getName() + "</b> is filtered out!", e.getProject());
         }
     }