+++ src/main/java/org/wavescale/sourcesync/config/FTPSConfiguration.java
@@ -22,7 +22,7 @@ public class FTPSConfiguration extends ConnectionConfiguration {
         super(connectionName);
         this.connectionType = ConnectionConstants.CONN_TYPE_FTPS;
         this.port = 21;
-        this.projectBasePath = "";
+        this.workspaceBasePath = "";
         this.host = "ftp://";
         this.userName = "";
         this.userPassword = "";