+++ src/main/java/org/wavescale/sourcesync/synchronizer/FTPSFileSynchronizer.java
@@ -80,7 +80,7 @@ public class FTPSFileSynchronizer extends FileSynchronizer {
         boolean preserveTimestamp = true;
         Path sourcePathLocation = Paths.get(sourceLocation);
         String sourceFileName = sourcePathLocation.getFileName().toString();
-        Path remotePath = Paths.get(this.getConnectionInfo().getProjectBasePath()).resolve(uploadLocation);
+        Path remotePath = Paths.get(this.getConnectionInfo().getWorkspaceBasePath()).resolve(uploadLocation);
 
         // first try to create the path where this must be uploaded
         try {