+++ CHANGELOG.md
@@ -3,10 +3,13 @@
 # Source Synchronizer Changelog
 
 ## [Unreleased]
+### Fixed
+- upload issues when using SFTP from local Windows to remote Linux
+- behavior for SSH keys component, now they properly enable or disable if users want to authenticate with SSH keys
 
 ### Changed
 
-- project's base location label from "Root path" to "Project base path". This is a breaking change, users
+- project's base location label from "Root path" to "Workspace base path". This is a breaking change, users
   will have to reconfigure the connections again.
 
 ## [2.0.0]